+++ src/include/LSP/LuauExt.hpp
@@ -33,8 +33,13 @@ using NameOrExpr = std::variant<std::string, Luau::AstExpr*>;
 
 // Converts a FTV and function call to a nice string
 // In the format "function NAME(args): ret"
-std::string toStringNamedFunction(
-    Luau::ModulePtr module, const Luau::FunctionTypeVar* ftv, const NameOrExpr nameOrFuncExpr, std::optional<Luau::ScopePtr> scope = std::nullopt);
+struct ToStringNamedFunctionOpts
+{
+    bool hideTableKind = false;
+};
+
+std::string toStringNamedFunction(Luau::ModulePtr module, const Luau::FunctionTypeVar* ftv, const NameOrExpr nameOrFuncExpr,
+    std::optional<Luau::ScopePtr> scope = std::nullopt, ToStringNamedFunctionOpts opts = {});
 
 std::string toStringReturnType(Luau::TypePackId retTypes, Luau::ToStringOptions options = {});
 Luau::ToStringResult toStringReturnTypeDetailed(Luau::TypePackId retTypes, Luau::ToStringOptions options = {});